+++ bot/modules/bartek/text_handler.py
